# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bina Academy API",
    "description": "## نظرة عامة | Overview\n\nمجموعة الـ APIs الخاصة بمنصة **Bina Academy** لطلبة التوجيهي (الفرع العلمي والأدبي - الصف 11 و 12).\n\nAPI collection for **Bina Academy** tutoring platform (Tawjihi students - grades 11/12, scientific/literary branches).\n\n---\n\n### خصائص عامة | General Properties\n\n- جميع الردود JSON (`Content-Type: application/json`)\n- جميع أرقام الهواتف بالصيغة الفلسطينية `^05\\d{8}$` (مثال: `0591234567`)\n- التواريخ بصيغة ISO 8601\n- المصادقة بـ JWT Bearer (توكينات الطلاب 30 يوم، توكينات الإدمن 7 أيام)\n- الحد العام للطلبات: 100 طلب/دقيقة لكل IP (حسب `@nestjs/throttler`)\n\n### تدفّق العمل الموصى به | Recommended Mobile Flows\n\n1. **تسجيل طالب جديد**: POST `/auth/register` (الهاتف، كلمة المرور، الاسم الكامل، الصف، الفرع) → تصفح المواد مباشرة.\n2. **دخول طالب مسجل**: POST `/auth/login` (الهاتف وكلمة المرور) → تصفح المواد.\n3. **مشاهدة فيديو**: GET `/subjects/:id/videos` → GET `/videos/:id/stream` → POST `/videos/:id/mark-viewed`\n4. **إدمن**: POST `/auth/admin/login` → استعمل الـ token في كل endpoints الـ admin\n",
    "version": "1.0.0",
    "contact": {
      "name": "Bina Academy Backend Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    },
    {
      "url": "http://10.0.2.2:3000/api",
      "description": "Android emulator (loopback to host)"
    },
    {
...
    }
  ],
...
  "paths": {
...
    "/subjects": {
      "get": {
        "tags": ["Subjects"],
        "summary": "قائمة المواد للطالب الحالي (حسب صفه وفرعه)",
        "description": "يقرأ `grade` و `branch` من ملف الطالب، ثم يعيد المواد المطابقة فقط مع عدد الفيديوهات المنشورة واسم المعلم. إذا لم يكن قد أكمل الملف الشخصي بعد، يرد بـ 404.\n\nReads the student's `grade` and `branch` from profile and returns matching subjects with published video count and teacher name. Returns 404 if profile incomplete.",
        "operationId": "listSubjects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة المواد",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SubjectListItem" }
                },
                "example": [
                  {
                    "id": "sub_math",
                    "name": "الرياضيات",
                    "grade": "GRADE_12",
                    "branch": "SCIENTIFIC",
                    "teacherName": "أ. خالد عمر",
                    "videoCount": 24,
                    "progressPercent": 75,
                    "locked": true
                  },
                  {
                    "id": "sub_phys",
                    "name": "الفيزياء",
                    "grade": "GRADE_12",
                    "branch": "SCIENTIFIC",
                    "teacherName": "أ. سامي حسن",
                    "videoCount": 18,
                    "progressPercent": 40,
                    "locked": false
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "الطالب لم يكمل الملف الشخصي بعد (لا grade/branch)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "يرجى إكمال الملف الشخصي أولاً"
                }
              }
            }
          }
        }
      }
    },
...
    "/subjects/{id}/videos": {
      "get": {
        "tags": ["Subjects"],
        "summary": "فيديوهات مادة معيّنة مع حالة القفل",
        "description": "يعرض كل الفيديوهات **المنشورة** للمادة مرتّبة حسب (unitNumber, orderInUnit). كل فيديو يحمل `locked: boolean` — يعتمد على `quota = plan.videosPerSubject` و **الترتيب** (`locked = (index >= quota)`)، وليس على مشاهدات سابقة.\n\n> ⚠️ **قاعدة الحصة**: لو الخطة تسمح بـ 10 فيديوهات، فالأول 10 فيديوهات فقط مفتوحة والباقي `locked: true` — حتى لو لم تُشاهد بعد.\n\nLists PUBLISHED videos in display order (unitNumber, orderInUnit). Each video has `locked: boolean` based on quota (top N unlocked). Locking is determined by **position in the sorted list**, not view history.",
        "operationId": "getSubjectVideos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "معرّف المادة",
            "schema": { "type": "string", "example": "sub_math" }
          }
        ],
        "responses": {
          "200": {
            "description": "المادة وفيديوهاتها وحصتها",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubjectVideosResponse" },
                "example": {
                  "subject": { "id": "sub_math", "name": "الرياضيات" },
                  "quota": 10,
                  "videos": [
                    {
                      "id": "vid_001",
                      "title": "النهايات - مقدمة",
                      "description": "تمهيد لمفهوم النهاية",
                      "durationSec": 645,
                      "unitNumber": 1,
                      "orderInUnit": 1,
                      "locked": false
                    },
                    {
                      "id": "vid_002",
                      "title": "النهايات - أمثلة",
                      "durationSec": 720,
                      "unitNumber": 1,
                      "orderInUnit": 2,
                      "locked": false
                    },
                    {
                      "id": "vid_011",
                      "title": "التفاضل - الوحدة 2",
                      "durationSec": 900,
                      "unitNumber": 2,
                      "orderInUnit": 1,
                      "locked": true
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "المادة غير موجودة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "statusCode": 404,
                  "message": "المادة غير موجودة"
                }
              }
            }
          }
        }
      }
    },
    "/videos/{id}": {
      "get": {
        "tags": ["Videos"],
        "summary": "تفاصيل الدرس وال chapters والفيديوهات المقترحة | Get Lesson Details",
        "description": "يعيد تفاصيل الدرس (الفيديو)، بما في ذلك رابط الـ PDF، معرف التقييم اليومي، المعلم، الفصول/القطع (chapters) مع حالاتها (منجزة، قيد التشغيل، مقفلة)، الفيديوهات ذات الصلة من نفس المادة، والعلامة المائية للطالب.",
        "operationId": "getLesson",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "بيانات الدرس بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "video": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "title": { "type": "string" },
                        "description": { "type": "string", "nullable": true },
                        "streamUrl": { "type": "string", "nullable": true },
                        "pdfUrl": { "type": "string", "nullable": true },
                        "durationSec": { "type": "integer" },
                        "unitName": { "type": "string" },
                        "subHeader": { "type": "string" },
                        "teacherName": { "type": "string", "nullable": true },
                        "dailyQuizId": { "type": "string", "nullable": true }
                      }
                    },
                    "chapters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "title": { "type": "string" },
                          "startSec": { "type": "integer" },
                          "endSec": { "type": "integer" },
                          "durationText": { "type": "string" },
                          "status": { "type": "string", "enum": ["COMPLETED", "PLAYING", "UNPLAYED", "LOCKED"] }
                        }
                      }
                    },
                    "relatedVideos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "title": { "type": "string" },
                          "subjectName": { "type": "string" },
                          "teacherName": { "type": "string", "nullable": true },
                          "durationText": { "type": "string" },
                          "unitName": { "type": "string" }
                        }
                      }
                    },
                    "watermark": {
                      "type": "object",
                      "properties": {
                        "name": { "type": "string" },
                        "phone": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "الفيديو غير موجود" }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "tags": ["Users"],
        "summary": "بيانات شاشة الرئيسية للموبايل | Mobile Home Screen Dashboard",
        "description": "يعيد بيانات لوحة معلومات الطالب بما في ذلك تقدمه العام، تفاصيل الفيديو الأخير لمتابعة التعلم، أهداف اليوم مع النسبة والرسالة الترحيبية، حالة الاختبار اليومي، والمواد المقترحة لليوم.",
        "operationId": "getDashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "بيانات لوحة المعلومات بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "studentName": { "type": "string" },
                    "generalProgress": { "type": "integer" },
                    "continueLearning": {
                      "type": "object",
                      "nullable": true,
                      "properties": {
                        "videoId": { "type": "string" },
                        "videoTitle": { "type": "string" },
                        "subjectName": { "type": "string" },
                        "unitName": { "type": "string" },
                        "lessonText": { "type": "string" },
                        "durationSec": { "type": "integer" },
                        "timeLeftMin": { "type": "integer" },
                        "progressPercent": { "type": "integer" }
                      }
                    },
                    "todayGoals": {
                      "type": "object",
                      "properties": {
                        "completedCount": { "type": "integer" },
                        "totalCount": { "type": "integer" },
                        "percentage": { "type": "integer" },
                        "text": { "type": "string" }
                      }
                    },
                    "dailyQuiz": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "title": { "type": "string" },
                        "description": { "type": "string" },
                        "isAvailable": { "type": "boolean" },
                        "buttonText": { "type": "string" },
                        "points": { "type": "integer" }
                      }
                    },
                    "suggestedSubjects": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "subjectName": { "type": "string" },
                          "teacherName": { "type": "string" },
                          "rating": { "type": "number" }
                        }
                      }
                    },
                    "subjectShopping": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "priceIls": { "type": "number" },
                          "isSubscribed": { "type": "boolean" },
                          "isInCart": { "type": "boolean" },
                          "teacher": {
                            "type": "object",
                            "nullable": true,
                            "properties": {
                              "id": { "type": "string" },
                              "name": { "type": "string" },
                              "avatarUrl": { "type": "string", "format": "uri" },
                              "rating": { "type": "number" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/admin/dashboard": {
      "get": {
        "tags": ["Admin"],
        "summary": "إحصائيات لوحة الإدارة",
        "description": "يعرض KPIs: عدد الطلاب الكلي/الفعّال/التجريبي، اشتراكات هذا الشهر والشهر السابق، إيراد تقديري بالشيكل، إجمالي الفيديوهات المنشورة ونسبة الإكمال، آخر 10 طلاب وأعلى 5 فيديوهات مشاهدةً. **يتطلب توكين ADMIN**.\n\nDashboard KPIs: students breakdown, monthly subscriptions, estimated revenue (ILS), published videos and completion rate, latest 10 students and top 5 videos by views.",
        "operationId": "getAdminDashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "إحصائيات شاملة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminDashboardStats" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/students": {
      "get": {
        "tags": ["Admin"],
        "summary": "قائمة الطلاب مع فلاتر",
        "description": "يعرض كل الطلاب (role=STUDENT) مع آخر اشتراك فعّال وعدّاد مشاهداتهم. يدعم فلاتر `grade`, `branch`, `search` (يبحث في `phone` و `name`)، و `subscriptionStatus`. **ملاحظة:** الفلتر `subscriptionStatus` مُعرَّف في DTO لكن **غير منفّذ في الـ query** حالياً — كل الفلترة الفعلية على grade/branch/search. يتطلب توكين ADMIN.\n\nList students with last active subscription and view count. Supports grade/branch/search filters. `subscriptionStatus` is defined in DTO but **not implemented in the query logic** — actual filtering works on grade/branch/search only.",
        "operationId": "listStudents",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "grade", "in": "query", "schema": { "type": "string", "enum": ["GRADE_11", "GRADE_12"] } },
          { "name": "branch", "in": "query", "schema": { "type": "string", "enum": ["SCIENTIFIC", "LITERARY"] } },
          { "name": "search", "in": "query", "description": "يبحث في phone و name (case-insensitive)", "schema": { "type": "string" } },
          { "name": "subscriptionStatus", "in": "query", "description": "⚠️ مستقبلي - غير منفّذ حالياً", "schema": { "type": "string", "enum": ["active", "trial", "expired", "none"] } }
        ],
        "responses": {
          "200": {
            "description": "قائمة الطلاب",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/StudentListItem" }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/students/{id}": {
      "get": {
        "tags": ["Admin"],
        "summary": "تفاصيل طالب واحد",
        "description": "يعرض طالباً مع كل اشتراكاته (مُرتّبة من الأحدث) وآخر 20 مشاهدة فيديو. يتطلب توكين ADMIN.",
        "operationId": "getStudent",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "بيانات الطالب الكاملة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "الطالب غير موجود",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "statusCode": 404, "message": "الطالب غير موجود" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Admin"],
        "summary": "تحديث بيانات طالب",
        "description": "يحدّث `grade`, `branch`, `parentPhone`, `isActive`. كل الحقول اختيارية. يتطلب توكين ADMIN.",
        "operationId": "updateStudent",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateStudentDto" },
              "example": { "grade": "GRADE_12", "branch": "SCIENTIFIC", "parentPhone": "0599876543", "isActive": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "الطالب بعد التحديث",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentListItem" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "الطالب غير موجود",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/admin/students/{id}/subscription/freeze": {
      "post": {
        "tags": ["Admin"],
        "summary": "تجميد/إلغاء تجميد اشتراك طالب",
        "description": "يضبط `isFrozen` على الـ subscription الفعّال الحالي. يستخدم عند السفر أو الانقطاع المؤقت. يتطلب توكين ADMIN.",
        "operationId": "freezeStudentSubscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FreezeSubscriptionDto" },
              "example": { "freeze": true, "reason": "سفر لمدة شهر" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم تجميد/إلغاء تجميد الاشتراك",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentSubscription" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "لا يوجد اشتراك فعّال",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "statusCode": 404, "message": "لا يوجد اشتراك فعّال" }
              }
            }
          }
        }
      }
    },
    "/admin/students/{id}/subscription/grant": {
      "post": {
        "tags": ["Admin"],
        "summary": "منح/تجديد اشتراك لطالب",
        "description": "يلغي أي اشتراك فعّال ويولّد اشتراكاً جديداً بنوع الخطة المحدّد. اختيارياً `durationDays` لتجاوز مدة الخطة الافتراضية. يتطلب توكين ADMIN.",
        "operationId": "grantStudentSubscription",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GrantSubscriptionDto" },
              "examples": {
                "monthly": { "summary": "اشتراك شهري", "value": { "planType": "MONTHLY" } },
                "trialCustom": { "summary": "تجريبي لمدة 14 يوم", "value": { "planType": "TRIAL", "durationDays": 14 } }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم منح الاشتراك",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StudentSubscription" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "خطة الاشتراك غير موجودة",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "statusCode": 404, "message": "الخطة غير موجودة" }
              }
            }
          }
        }
      }
    },
    "/admin/videos": {
      "get": {
        "tags": ["Admin"],
        "summary": "قائمة كل الفيديوهات (admin)",
        "description": "كل الفيديوهات بكل الحالات (DRAFT/PENDING_REVIEW/PUBLISHED) مع المادة والمعلم وعدّاد المشاهدات. يتطلب توكين ADMIN.",
        "operationId": "listAdminVideos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة الفيديوهات",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AdminVideoListItem" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "إنشاء فيديو جديد",
        "description": "ينشئ فيديو جديد بحالة DRAFT افتراضياً. يتطلب توكين ADMIN.",
        "operationId": "createAdminVideo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVideoDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "تم إنشاء الفيديو",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Video" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/videos/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "تحديث فيديو",
        "description": "PATCH-style update — كل الحقول في DTO اختيارية. يتطلب توكين ADMIN.",
        "operationId": "updateAdminVideo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateVideoDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "الفيديو بعد التحديث",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Video" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "حذف فيديو (admin - hard delete)",
        "description": "**تحذير**: حذف نهائي — يحذف الـ Video وكل VideoViews المرتبطة (CASCADE). يتطلب توكين ADMIN.",
        "operationId": "deleteAdminVideo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "تم الحذف",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Video" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/subjects": {
      "get": {
        "tags": ["Admin"],
        "summary": "قائمة كل المواد (admin)",
        "description": "كل المواد مرتّبة بالصف/الفرع/الاسم مع اسم المعلم وعدّاد الفيديوهات. يتطلب توكين ADMIN.",
        "operationId": "listAdminSubjects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة المواد",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Subject" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/plans": {
      "get": {
        "tags": ["Admin"],
        "summary": "قائمة كل خطط الاشتراك (حتى غير الفعّالة)",
        "description": "كل الخطط مرتّبة بالمدة. يتطلب توكين ADMIN.",
        "operationId": "listAdminPlans",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "قائمة الخطط",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SubscriptionPlan" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/admin/plans/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "تحديث خطة اشتراك",
        "description": "PATCH-style — `priceIls`, `videosPerSubject`, `discountPercent`, `isActive`. يتطلب توكين ADMIN.",
        "operationId": "updateAdminPlan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePlanDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "الخطة بعد التحديث",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubscriptionPlan" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/admin/teachers/dashboard": {
      "get": {
        "tags": ["Admin"],
        "summary": "إحصائيات وطلبات المعلمين | Teachers Dashboard Stats & Jobs",
        "description": "يعود بإحصائيات المعلمين، طلبات التوظيف الجديدة وقائمة المعلمين المتميزين. يتطلب توكين ADMIN.\n\nReturns teachers KPI stats, new job applications, and top teachers list. Requires ADMIN JWT.",
        "operationId": "getTeachersDashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "بيانات لوحة المعلمين بنجاح",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stats": {
                      "type": "object",
                      "properties": {
                        "totalTeachers": { "type": "integer" },
                        "activeClasses": { "type": "integer" },
                        "performanceRating": { "type": "number" },
                        "contentHours": { "type": "integer" }
                      }
                    },
                    "applications": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "title": { "type": "string" },
                          "timeText": { "type": "string" }
                        }
                      }
                    },
                    "topTeachers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "satisfactionRate": { "type": "integer" },
                          "avatar": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer.\n\n- توكين STUDENT: يولَّد من `/auth/verify-otp` (صالح 30 يوم، يستخدم `JWT_SECRET`)\n- توكين ADMIN: يولَّد من `/auth/admin/login` (صالح 7 أيام، يستخدم `ADMIN_JWT_SECRET` أو `JWT_SECRET` كاحتياط)\n\n> ⚠️ **مهم**: لا تستخدم توكين STUDENT في endpoints الـ Admin والعكس صحيح — كلاهما سيرد بـ 401."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "التوكن مفقود/منتهي/غير صالح أو الـ role لا يطابق الـ endpoint",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "examples": {
              "missingToken": { "summary": "بدون توكن", "value": { "statusCode": 401, "message": "Unauthorized" } },
              "badRole": { "summary": "توكين لا يطابق الـ role", "value": { "statusCode": 401, "message": "جلسة غير صالحة" } }
            }
          }
        }
      }
    },
...
  }
}