# api/src/admin/admin.controller.ts
import {
  Controller,
  Get,
  Post,
  Put,
  Delete,
  Param,
  Body,
  Query,
  UseGuards,
  UseInterceptors,
  UploadedFile,
  BadRequestException,
} from '@nestjs/common';
import { FileInterceptor } from '@nestjs/platform-express';
import { diskStorage } from 'multer';
import { extname } from 'path';
import { AdminService } from './admin.service';
import { AdminJwtGuard } from '../auth/guards/jwt.guard';
import {
  ListStudentsQuery,
  UpdateStudentDto,
  FreezeSubscriptionDto,
  GrantSubscriptionDto,
  CreateVideoDto,
  UpdatePlanDto,
  UpdateSubjectDto,
  CreateTeacherDto,
} from './dto/admin.dto';

@Controller('admin')
@UseGuards(AdminJwtGuard)
export class AdminController {
  constructor(private adminService: AdminService) {}

  @Get('dashboard')
  getDashboard() {
    return this.adminService.getDashboardStats();
  }

  @Get('students')
  listStudents(@Query() query: ListStudentsQuery) {
    return this.adminService.listStudents(query);
  }

  @Get('students/:id')
  getStudent(@Param('id') id: string) {
    return this.adminService.getStudent(id);
  }

  @Put('students/:id')
  updateStudent(@Param('id') id: string, @Body() dto: UpdateStudentDto) {
    return this.adminService.updateStudent(id, dto);
  }

  @Post('students/:id/subscription/freeze')
  freezeSubscription(@Param('id') id: string, @Body() dto: FreezeSubscriptionDto) {
    return this.adminService.freezeSubscription(id, dto);
  }

  @Post('students/:id/subscription/grant')
  grantSubscription(@Param('id') id: string, @Body() dto: GrantSubscriptionDto) {
    return this.adminService.grantSubscription(id, dto);
  }

  @Get('videos')
  listVideos() {
    return this.adminService.listVideos();
  }

  @Post('videos')
  createVideo(@Body() dto: CreateVideoDto) {
    return this.adminService.createVideo(dto);
  }

  @Put('videos/:id')
  updateVideo(@Param('id') id: string, @Body() dto: Partial<CreateVideoDto>) {
    return this.adminService.updateVideo(id, dto);
  }

  @Delete('videos/:id')
  deleteVideo(@Param('id') id: string) {
    return this.adminService.deleteVideo(id);
  }

  @Get('subjects')
  listSubjects() {
    return this.adminService.listSubjects();
  }

  @Put('subjects/:id')
  updateSubject(@Param('id') id: string, @Body() dto: UpdateSubjectDto) {
    return this.adminService.updateSubject(id, dto);
  }


  @Get('plans')
  listPlans() {
    return this.adminService.listPlans();
  }

  @Put('plans/:id')
  updatePlan(@Param('id') id: string, @Body() dto: UpdatePlanDto) {
    return this.adminService.updatePlan(id, dto);
  }

  @Get('teachers')
  listTeachers(@Query() query: { search?: string; page?: string; limit?: string }) {
    return this.adminService.listTeachers(query);
  }

  @Get('teachers/dashboard')
  getTeachersDashboard() {
    return this.adminService.getTeachersDashboard();
  }

  @Post('teachers')
  createTeacher(@Body() dto: CreateTeacherDto) {
    return this.adminService.createTeacher(dto);
  }

  @Post('upload')
  @UseInterceptors(
    FileInterceptor('file', {
      storage: diskStorage({
        destination: './uploads',
        filename: (req, file, callback) => {
          const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1e9);
          const ext = extname(file.originalname);
          callback(null, `${uniqueSuffix}${ext}`);
        },
      }),
    }),
  )
  uploadFile(@UploadedFile() file: any) {
    if (!file) {
      throw new BadRequestException('لم يتم تحديد أي ملف للرفع');
    }
    return {
      url: `/uploads/${file.filename}`,
    };
  }
}